Docs and restructuring
