+++ django_guid/integrations/celery/config.py
