+++ django_guid/integrations/celery/context.py
